refactored package structure on view components
