+++ src/org/uem/dam/GestorFarmacia/view/submenus/data_view/TableSubmenu.java
@@ -1,4 +1,4 @@
-package org.uem.dam.GestorFarmacia.view.submenus;
+package org.uem.dam.GestorFarmacia.view.submenus.data_view;
 
 import java.awt.event.ActionListener;
 import java.util.ArrayList;
@@ -9,6 +9,7 @@ import javax.swing.table.DefaultTableModel;
 
 import org.uem.dam.GestorFarmacia.model.DBItem;
 import org.uem.dam.GestorFarmacia.view.ComponentView;
+import org.uem.dam.GestorFarmacia.view.submenus.DefaultSubmenu;
 
 import net.miginfocom.swing.MigLayout;
